// update_app_01082026/Railmate_App-FIXED-final/Railmate_App-main-FIXED/apk/types/report.types.ts
export type ReportType =
  | 'DELAY'
  | 'CROWD'
  | 'PLATFORM'
  | 'SCHEDULE'
  | 'GENERAL'
  | 'ACCIDENT';